+++ src/components/themes/models/FriendDetail.ts
@@ -1,0 +1,3 @@
+export interface FriendDetailProps {
+  friendId: string | undefined;
+}
