+++ src/components/themes/models/MessageRoom.ts
@@ -1,0 +1,3 @@
+export interface MessageRoomProps {
+  messageRoomId: string | undefined;
+}
