+++ src/components/themes/models/MessageRoomSettings.ts
@@ -1,0 +1,3 @@
+import { MessageRoomProps } from "@/components/themes/models/MessageRoom.ts";
+
+export type MessageRoomSettingsProps = MessageRoomProps;
