+++ src/components/themes/models/Settings.ts
@@ -1,0 +1,5 @@
+import { GlobalConfig } from "@/domain/config/models.ts";
+
+export type SettingsProps = {
+  config: GlobalConfig;
+};
