+++ src/components/themes/theFruit/SettingSponsor.tsx
@@ -1,0 +1,9 @@
+import SettingTopBar from "@/components/common/SettingTopBar.tsx";
+
+export function SettingSponsor() {
+  return (
+    <div className="size-full bg-emptyBackground">
+      <SettingTopBar title={"Sponsors"} enableHome />
+    </div>
+  );
+}
