+++ src/components/themes/theFruit/TypingSimulationSelector.tsx
