+++ src/routes/main/friends/[:friendId]/settings/conversation-style/page.tsx
@@ -2,12 +2,17 @@ import { useRightPrimaryPage } from "@/routes/main/hooks.ts";
 import { useDb } from "@/contexts/DbContext.ts";
 import { useLiveQuery } from "dexie-react-hooks";
 import { makeGlobalConfigRepository } from "@/domain/config/repository.ts";
-import ConversationStyleSelector from "@/routes/main/friends/[:friendId]/settings/conversation-style/ConversationStyleSelector.tsx";
+import { useTheme } from "@/hooks/useTheme.ts";
 
 export function Component() {
   useRightPrimaryPage();
   const db = useDb();
   const config = useLiveQuery(makeGlobalConfigRepository(db).getGlobalConfig);
+  const theme = useTheme();
+  if (!theme) return null;
+  const {
+    components: { ConversationStyleSelector },
+  } = theme;
 
   return <>{config && <ConversationStyleSelector config={config} />}</>;
 }
