+++ src/routes/main/friends/[:friendId]/settings/page.tsx
@@ -2,13 +2,18 @@ import { useRightPrimaryPage } from "@/routes/main/hooks.ts";
 import { useDb } from "@/contexts/DbContext.ts";
 import { useLiveQuery } from "dexie-react-hooks";
 import { makeGlobalConfigRepository } from "@/domain/config/repository.ts";
-import SettingsPage from "@/routes/main/friends/[:friendId]/settings/SettingsPage.tsx";
+import { useTheme } from "@/hooks/useTheme.ts";
 
 export function Component() {
   useRightPrimaryPage();
   const db = useDb();
   const config = useLiveQuery(makeGlobalConfigRepository(db).getGlobalConfig);
-  return <>{config && <SettingsPage config={config} />}</>;
+  const theme = useTheme();
+  if (!theme) return null;
+  const {
+    components: { FriendSettings },
+  } = theme;
+  return <>{config && <FriendSettings config={config} />}</>;
 }
 
 Component.displayName = "FriendSettingsPage";
