+++ src/routes/main/settings/chat-customize/persona/page.tsx
@@ -1,16 +1,19 @@
-import { useTranslation } from "react-i18next";
-import SettingTopBar from "@/components/common/SettingTopBar.tsx";
 import { useRightPrimaryPage } from "@/routes/main/hooks.ts";
+import { useTheme } from "@/hooks/useTheme.ts";
+import { useDb } from "@/contexts/DbContext.ts";
+import { useLiveQuery } from "dexie-react-hooks";
+import { makeGlobalConfigRepository } from "@/domain/config/repository.ts";
 
 export function Component() {
   useRightPrimaryPage();
-  const { t } = useTranslation("pages/friends");
-  return (
-    <div className="size-full bg-emptyBackground">
-      <SettingTopBar title={t("settings.options.self-intro")} enableBack />
-      <div className="flex w-full flex-col items-center px-[15px] tablet:px-[80px] desktop:px-[190px]"></div>
-    </div>
-  );
-}
+  const db = useDb();
+  const config = useLiveQuery(makeGlobalConfigRepository(db).getGlobalConfig);
+  const theme = useTheme();
+  if (!theme || !config) return null;
+  const {
+    components: { PersonalSettings },
+  } = theme;
 
+  return <PersonalSettings config={config} />;
+}
 Component.displayName = "PersonaGlobalSettingPage";
